pyproject.toml
[project]
name = "conference-agent"
version = "0.1.0"
description = "컨퍼런스 아젠다 리서치 및 프로그램 기획 AI 에이전트"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.44",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "streamlit>=1.40.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "semgrep",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
